% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{clique.community}
\alias{clique.community}
\title{This function detects communities using cliques}
\usage{
clique.community(graph, k)
}
\arguments{
\item{graph}{is igraph object}

\item{k}{is the number of communitites to find}
}
\description{
This function detects communities using cliques
}
\references{
Palla, Gergely, et al. "Uncovering the overlapping community structure of complex networks in nature and society." Nature 435.7043 (2005): 814-818.
}

