% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{community.significance.test}
\alias{community.significance.test}
\title{This function performs a Wilcoxon rank-sum test on the "internal" and "external" degrees of a community in order to quantify its significance.}
\usage{
community.significance.test(graph, vs, ...)
}
\description{
The edges within a community are "internal" and the edges connecting the vertices of a community with the rest of the graph are "external". More internal than external edges show that the community is significant; the otherwise suggests that the community is in fact an "anti-community".
}

