% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{community_influence}
\alias{community_influence}
\title{This function finds influential nodes in communities in given graph}
\usage{
community_influence()
}
\description{
This function finds influential nodes in communities in given graph
}

