% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{find_communities}
\alias{find_communities}
\title{This method finds communities in the given graph and returns the graph after adding a vector "group" to its vertices}
\usage{
find_communities(graph, plot = TRUE, method = c("multilevel",
  "edgebetweenness", "fastgreedy", "eigenvector", "spinglass", "walktrap",
  "labelpropagation", "clique", "largescale"))
}
\arguments{
\item{graph}{is the igraph object}

\item{plot}{whether to plot the graph using tkplot. Default is TRUE}

\item{method}{is the method to find communities. Value can be "multilevel", "edgebetweenness", "fastgreedy", "eigenvector", "spinglass", "walktrap", "labelpropagation", "clique", "largescale"}

\item{prob}{is the probability of activation of a neighbour node. This is applicable only to IC model currently}
}
\description{
This method finds communities in the given graph and returns the graph after adding a vector "group" to its vertices
}

