% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{graph_summary}
\alias{graph_summary}
\title{This functions gives a summary of common metrics of given graph}
\usage{
graph_summary(graph, plot = FALSE)
}
\arguments{
\item{graph}{is the igraph object}

\item{plot}{uses tkplot to plot the graph. Default is FALSE}
}
\value{
object containing summary
}
\description{
This functions gives a summary of common metrics of given graph
}

