% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{influence}
\alias{influence}
\title{This function is a wrapper for influence_IC and influence_LT functions}
\usage{
influence(graph, seed = NULL, budget = 1, steps = 1, model = c("IC",
  "LT"), maximize = FALSE, seed_method = c("random", "degree", "closeness",
  "betweenness", "coreness", "eigenvector", "a-degree", "a-closeness",
  "a-betweenness", "a-coreness", "a-eigenvector"), prob = 0.5)
}
\arguments{
\item{graph}{is the igraph object}

\item{seed}{(optional) is a set of seed (initial nodes). If NULL, then seed_method parameter should be given}

\item{budget}{defines what percentage of most influential nodes out of all nodes is required as output. Default value is 1}

\item{steps}{is the time steps for which, the diffusion process should run. If exhaustive run is required, provide a high value (like 100). Default value is 1}

\item{model}{is influence model to run the dataset on. Value MUST either be "LT" or "IC"}

\item{maximize}{should be TRUE if influential nodes are to be derived using Greedy algorithm}

\item{seed_method}{is the selection method for seed (initial nodes). Value can be "random", "degree", "closeness", "betweenness", "coreness", "eigenvector", "a-degree", "a-closeness", "a-betweenness", "a-coreness", "a-eigenvector"}

\item{prob}{is the probability of activation of a neighbour node. This is applicable only to IC model currently}
}
\value{
output containing summary
}
\description{
This function is a wrapper for influence_IC and influence_LT functions
}
\examples{
influence(G, budget=5, seed=NULL, 5, "LT", maximize=TRUE, seed_method="degree", prob=0.5)
influence(G, budget=5, seed=NULL, 5, "IC", maximize=TRUE, seed_method="degree", prob=0.5)
influence(G, budget=5, seed=c(2,5,9,23), 5, "IC", maximize=FALSE, prob=0.5)
}

