% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{large.scale.community}
\alias{large.scale.community}
\title{This function detects communities in large-scale graphs}
\usage{
large.scale.community(graph, mode = "all")
}
\arguments{
\item{graph}{is igraph object}

\item{mode}{is detection mode. Values can be "all", "in" or "out". Default is "all"}
}
\description{
This function detects communities in large-scale graphs
}
\references{
Raghavan, Usha Nandini, Réka Albert, and Soundar Kumara. "Near linear time algorithm to detect community structures in large-scale networks." Physical Review E 76.3 (2007): 036106.
}

