% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{largest_component}
\alias{largest_component}
\title{This function returns largest connected component in a network}
\usage{
largest_component(graph)
}
\arguments{
\item{graph}{is the igraph object}
}
\value{
largest component igraph object
}
\description{
This function returns largest connected component in a network
}

