% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{new_influence_max}
\alias{new_influence_max}
\title{Try to improve maximization. NOT TESTED}
\usage{
new_influence_max(G, budget, steps, model, prob)
}
\description{
Try to improve maximization. NOT TESTED
}

