% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{select_seed}
\alias{select_seed}
\title{This function inputs a graph object, percentage and a seed method and returns k% nodes as seed using given method}
\usage{
select_seed(graph, budget, seed_method = c("random", "degree", "closeness",
  "betweenness", "coreness", "eigenvector", "a-degree", "a-closeness",
  "a-betweenness", "a-coreness", "a-eigenvector"))
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{defines what percentage of most influential nodes out of all nodes is required as output. Default value is 1}

\item{seed_method}{is the selection method for seed (initial nodes). Value can be "random", "degree", "closeness", "betweenness", "coreness", "eigenvector", "a-degree", "a-closeness", "a-betweenness", "a-coreness", "a-eigenvector"}
}
\value{
set of nodes
}
\description{
This function inputs a graph object, percentage and a seed method and returns k% nodes as seed using given method
}

