% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Influence.R
\name{new_influence_max}
\alias{new_influence_max}
\title{Try to improve maximization. NOT TESTED}
\usage{
new_influence_max(graph, budget, steps, model, prob)
}
\arguments{
\item{graph}{is the igraph object}

\item{budget}{defines what percentage of most influential nodes out of all nodes is required as output. Default value is 1}

\item{steps}{is the time steps for which, the diffusion process should run. If exhaustive run is required, provide a high value (like 100). Default value is 1}

\item{model}{is influence model to run the dataset on. Value MUST either be "LT" or "IC"}

\item{prob}{is the probability of activation of a neighbour node. This is applicable only to IC model currently}
}
\value{
output summary
}
\description{
Try to improve maximization. NOT TESTED
}

